1)为基于Framework的系统减少依赖,根据编译条件,Framework 和.netCore 分别使用 WebClient和HttpClient.Framework 的Json序列化用 Newtonsoft.Json. 2)增加跳过当前版本的更新功能

+++ AttnSoft.AutoUpdate/Middlewares/UpgradMiddleware.cs
@@ -4,7 +4,7 @@ using System.Threading.Tasks;
 namespace AttnSoft.AutoUpdate.Middlewares;
 public interface IUpgrad : IApplicationMiddleware<UpdateContext> { }
 /// <summary>
-/// 升级自身的Upgrade
+/// Upgrade yourself.升级自身的Upgrade
 /// </summary>
 public class UpgradMiddleware : IUpgrad
 {
@@ -13,26 +13,51 @@ public class UpgradMiddleware : IUpgrad
         var sourcePath = context.AppPath;
         var patchPath = context.PatchPath;
         var upgradeName = context.UpgradName;
-        upgradeName = Path.GetFileNameWithoutExtension(upgradeName);
-        string[] upgradeFiles = new string[] {
+
+#if NETFRAMEWORK
+        string resUpgradeName = "AttnSoft.AutoUpdate.Upgrade.Framework.Upgrade.exe";
+#else
+        string resUpgradeName = "AttnSoft.AutoUpdate.Upgrade.AOT.Upgrade.exe";
+#endif
+        System.Reflection.Assembly app = System.Reflection.Assembly.GetExecutingAssembly();
+        Stream? fsOpen = app.GetManifestResourceStream(resUpgradeName);
+        if (fsOpen != null)
+        {
+            context.UpgradName = "Upgrade.exe";
+            var filename = Path.Combine(sourcePath, context.UpgradName);
+            FileStream fsSave = new FileStream(filename, FileMode.Create);
+            int readcount = 0;
+            byte[] bytes = new byte[1024];
+            while ((readcount = fsOpen.Read(bytes, 0, 1024)) > 0)
+            {
+                fsSave.Write(bytes, 0, readcount);
+            }
+            fsSave.Close();
+            fsOpen.Close();
+        }
+        else
+        {
+            upgradeName = Path.GetFileNameWithoutExtension(upgradeName);
+            string[] upgradeFiles = new string[] {
             upgradeName,
             $"{upgradeName}.exe",
             $"{upgradeName}.dll",
             $"{upgradeName}.deps.json",
             $"{upgradeName}.runtimeconfig.json"
             };
-        foreach (var upgradeFile in upgradeFiles)
-        {
-            var newUpgradeFile = Path.Combine(patchPath, upgradeFile);
-            if (File.Exists(newUpgradeFile))
+            foreach (var upgradeFile in upgradeFiles)
             {
-                var oldUpgradeFile = Path.Combine(sourcePath, upgradeFile);
-                if (File.Exists(oldUpgradeFile))
+                var newUpgradeFile = Path.Combine(patchPath, upgradeFile);
+                if (File.Exists(newUpgradeFile))
                 {
-                    File.SetAttributes(oldUpgradeFile, FileAttributes.Normal);
-                    File.Delete(oldUpgradeFile);
+                    var oldUpgradeFile = Path.Combine(sourcePath, upgradeFile);
+                    if (File.Exists(oldUpgradeFile))
+                    {
+                        File.SetAttributes(oldUpgradeFile, FileAttributes.Normal);
+                        File.Delete(oldUpgradeFile);
+                    }
+                    File.Move(newUpgradeFile, oldUpgradeFile);
                 }
-                File.Move(newUpgradeFile, oldUpgradeFile);
             }
         }
         await next(context);
